requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
discord.py==2.5.2
exceptiongroup==1.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
sniffio==1.3.1
typing_extensions==4.14.1
watchgod==0.8.2
yarl==1.20.1